TSQL: Finish all variants of expressions (#505)
- Implement DEFAULT as a primitive expression
- Implement @localvar as a primitive expression
- Provide coverage test
- Removes defaulting to visiting terminal nodes

Closes: #275

+++ core/src/test/scala/com/databricks/labs/remorph/parsers/tsql/TSqlRelationBuilderSpec.scala
@@ -47,7 +47,7 @@ class TSqlRelationBuilderSpec
       example(
         "FROM some_table WHERE 1=1",
         _.selectOptionalClauses(),
-        ir.Filter(namedTable("some_table"), ir.Equals(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(1)))))
+        ir.Filter(namedTable("some_table"), ir.Equals(ir.Literal(short = Some(1)), ir.Literal(short = Some(1)))))
     }
 
     "translate GROUP BY clauses" in {
@@ -91,9 +91,8 @@ class TSqlRelationBuilderSpec
         "FROM some_table WHERE 1=1 GROUP BY some_column",
         _.selectOptionalClauses(),
         ir.Aggregate(
-          input = ir.Filter(
-            namedTable("some_table"),
-            ir.Equals(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(1)))),
+          input =
+            ir.Filter(namedTable("some_table"), ir.Equals(ir.Literal(short = Some(1)), ir.Literal(short = Some(1)))),
           group_type = ir.GroupBy,
           grouping_expressions = Seq(simplyNamedColumn("some_column")),
           pivot = None))
@@ -103,9 +102,8 @@ class TSqlRelationBuilderSpec
         _.selectOptionalClauses(),
         ir.Sort(
           ir.Aggregate(
-            input = ir.Filter(
-              namedTable("some_table"),
-              ir.Equals(ir.Literal(integer = Some(1)), ir.Literal(integer = Some(1)))),
+            input =
+              ir.Filter(namedTable("some_table"), ir.Equals(ir.Literal(short = Some(1)), ir.Literal(short = Some(1)))),
             group_type = ir.GroupBy,
             grouping_expressions = Seq(simplyNamedColumn("some_column")),
             pivot = None),
